Added access control to files

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,17 @@ use App\Post;
 
 class PostController extends Controller
 {
+
+
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -79,6 +90,10 @@ class PostController extends Controller
     {
         //
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id != $post->user->id){
+            return redirect('/posts')->with('error','unaurthorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -115,6 +130,9 @@ class PostController extends Controller
     public function destroy($id) 
     {
         $post = Post::find($id);
+        if(auth()->user()->id != $post->user->id){
+            return redirect('/posts')->with('error','unaurthorized page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted successfully');
